Fix crash on book fetch with no internet
This commit refactors the book service to handle some error cases that
can arise when attempting to search for a book without an internet
connection. It also adds a small delay on the text change listener so
that we are not sending search queries before the user has finished
typing their ISBN number entirely.