Add next match widget
This commit adds a widget that will show the user when today's next
match will be played. It should address Jennie's user feedback about
having to open the app to check for an upcoming game. The widget will
update itself every 30 minutes to stay up to date as well as receiving
updates whenever the fetch task is run.