Fix NullPointer issues in BookDetail.java
This will fix two potential crash sites in the BookDetail.java file. The
first was related to a potentially null ShareActionProvider. The
activity will now attempt to setup the share intent only in the
ShareActionProvider has been initialized. The second issue was related
to a book without an author. The code had assumed that every book would
have an author and this would cause books without authors to crash the
app. The BookDetail activity will now check for an author and in the
absence of one it will just fill in "Unknown".