Eliminate duplicated author view binding logic
The NullPointer crashes related to books without authors was causing a
crash in two different activities because of some duplicated logic
related to pulling author information out of a cursor and putting it
into a textview. I chose to combine that logic into one place and put
it in a static method call rather than fix the bug in two different
places and continue the duplication.