Support RTL in all layouts
This should add RTL support to all of the apps layouts. I had missed a
few during my initial submission. I am also adding a layout-v17 layout
file because when I was adding support for the 'paddingStart' attribute
the tools detected it as an error that would cause problems for devices
below api 17 and the auto fix recommended simply performing an override
in an alternate layout file. I chose to do this option even though the
attributes for 'paddingEnd' and 'drawableStart' did not cause the same
error flag to be raised.